Start of original gui refactor - lose the interfaces and Generic* prefix for all gui classes, remove most depreciated methods, add some javadocs, and fix any local dependencies.
All changes are local only, the protocol is unchanged so client changes are separate and this won't affect it (they will be merged in there at a  later point).
