+++ src/main/java/org/getspout/spoutapi/gui/ComboBox.java
@@ -19,26 +19,144 @@
  */
 package org.getspout.spoutapi.gui;
 
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
-public interface ComboBox extends Button {
-	public ComboBox setItems(List<String> items);
+import org.getspout.spoutapi.io.SpoutInputStream;
+import org.getspout.spoutapi.io.SpoutOutputStream;
+import org.getspout.spoutapi.packet.PacketComboBox;
 
-	public List<String> getItems();
+public class ComboBox extends Button {
+	private List<String> items = new ArrayList<String>();
+	private boolean open = false;
+	private int selection = -1;
+	private String format = "%text%: %selected%";
 
-	public ComboBox openList();
+	/**
+	 * Set a list of options
+	 * @param items
+	 * @return
+	 */
+	public ComboBox setItems(List<String> items) {
+		this.items = items;
+		return this;
+	}
 
-	public ComboBox closeList();
+	/**
+	 * Get all options
+	 * @return
+	 */
+	public List<String> getItems() {
+		return items;
+	}
 
-	public String getSelectedItem();
+	/**
+	 * Open the combo box
+	 * @return
+	 */
+	public ComboBox openList() {
+		setOpen(true, true);
+		return this;
+	}
 
-	public int getSelectedRow();
+	/**
+	 * Close the combo box
+	 * @return
+	 */
+	public ComboBox closeList() {
+		setOpen(false, true);
+		return null;
+	}
 
-	public ComboBox setSelection(int row);
+	/**
+	 * Get the label from the selected option
+	 * @return
+	 */
+	public String getSelectedItem() {
+		if (selection >= 0 && selection < items.size()) {
+			return items.get(selection);
+		} else {
+			return null;
+		}
+	}
 
-	public void onSelectionChanged(int i, String text);
+	/**
+	 * Get which option is selected
+	 * @return
+	 */
+	public int getSelectedRow() {
+		return this.selection;
+	}
 
-	public boolean isOpen();
+	/**
+	 * Set which option is active
+	 * @param row
+	 * @return
+	 */
+	public ComboBox setSelection(int row) {
+		boolean event = row != selection;
+		this.selection = row;
+		if (event) {
+			onSelectionChanged(row, getSelectedItem());
+		}
+		return this;
+	}
+
+	public void onSelectionChanged(int i, String text) {
+	}
+
+	public boolean isOpen() {
+		return open;
+	}
+
+	/**
+	 * Sets the open status.
+	 * @param open       the state
+	 * @param sendPacket if true, send an update packet
+	 * @return the instance
+	 */
+	public ComboBox setOpen(boolean open, boolean sendPacket) {
+		if (sendPacket) {
+			if (open != this.open) {
+				this.open = open;
+				PacketComboBox packet = new PacketComboBox(this);
+				getScreen().getPlayer().sendPacket(packet);
+			}
+		}
+		this.open = open;
+		return this;
+	}
+
+	@Override
+	public WidgetType getType() {
+		return WidgetType.ComboBox;
+	}
+
+	@Override
+	public void readData(SpoutInputStream input) throws IOException {
+		super.readData(input);
+	}
+
+	@Override
+	public void writeData(SpoutOutputStream output) throws IOException {
+		super.writeData(output);
+		output.writeInt(selection);
+		output.writeInt(getItems().size());
+		for (String item : getItems()) {
+			output.writeString(item);
+		}
+		output.writeString(format);
+	}
+
+	@Override
+	public int getVersion() {
+		return super.getVersion() + 1;
+	}
+
+	public String getFormat() {
+		return format;
+	}
 
 	/**
 	 * Sets the format of the text on the button. Default is "%text%: %selected%"
@@ -48,7 +166,18 @@ public interface ComboBox extends Button {
 	 * @param format the format of the text on the button
 	 * @return the instance
 	 */
-	public ComboBox setFormat(String format);
+	public ComboBox setFormat(String format) {
+		this.format = format;
+		return this;
+	}
 
-	public String getFormat();
+	@Override
+	public String getText() {
+		if (super.getText() == null || super.getText().isEmpty()) {
+			return getSelectedItem();
+		} else {
+			String text = format.replaceAll("%text%", super.getText()).replaceAll("%selected%", getSelectedItem());
+			return text;
+		}
+	}
 }
