+++ src/main/java/org/getspout/spoutapi/gui/ItemWidget.java
@@ -19,6 +19,13 @@
  */
 package org.getspout.spoutapi.gui;
 
+import java.io.IOException;
+
+import org.bukkit.inventory.ItemStack;
+
+import org.getspout.spoutapi.io.SpoutInputStream;
+import org.getspout.spoutapi.io.SpoutOutputStream;
+
 /**
  * The GenericItemWidget class allows you to display a block or item as it
  * would be in the player's inventory.
@@ -26,49 +33,122 @@ package org.getspout.spoutapi.gui;
  * Don't forget that most items are in fact 3d, so also need a depth to draw
  * properly.
  */
-public interface ItemWidget extends Widget {
+public class ItemWidget extends Widget {
+	protected int material = -1;
+	protected short data = -1;
+	protected int depth = 8;
+
+	public ItemWidget() {
+	}
+
+	public ItemWidget(ItemStack item) {
+		this.material = item.getTypeId();
+		this.data = item.getDurability();
+	}
+
+	@Override
+	public int getVersion() {
+		return super.getVersion() + 0;
+	}
+
+	@Override
+	public void readData(SpoutInputStream input) throws IOException {
+		super.readData(input);
+		this.setTypeId(input.readInt());
+		this.setData(input.readShort());
+		this.setDepth(input.readInt());
+	}
+
+	@Override
+	public void writeData(SpoutOutputStream output) throws IOException {
+		super.writeData(output);
+		output.writeInt(getTypeId());
+		output.writeShort(getData());
+		output.writeInt(getDepth());
+	}
+
 	/**
 	 * Sets the type id of this item widget
 	 * @param id
 	 * @return ItemWidget
 	 */
-	public ItemWidget setTypeId(int id);
+	public ItemWidget setTypeId(int id) {
+		if (getTypeId() != id) {
+			this.material = id;
+			autoDirty();
+		}
+		return this;
+	}
 
 	/**
 	 * Gets the type id of this item widget
 	 * @return type id
 	 */
-	public int getTypeId();
+	public int getTypeId() {
+		return material;
+	}
 
 	/**
 	 * Sets the data of this item widget
 	 * @param data to set
 	 * @return ItemWidget
 	 */
-	public ItemWidget setData(short data);
+	public ItemWidget setData(short data) {
+		if (getData() != data) {
+			this.data = data;
+			autoDirty();
+		}
+		return this;
+	}
 
 	/**
 	 * Gets the data of this item widget, is zero by default
 	 * @return data
 	 */
-	public short getData();
+	public short getData() {
+		return data;
+	}
 
 	/**
 	 * Sets the z render depth for this 3-d item widget
 	 * @param depth to render at
 	 * @return ItemWidget
 	 */
-	public ItemWidget setDepth(int depth);
+	public ItemWidget setDepth(int depth) {
+		if (getDepth() != depth) {
+			this.depth = depth;
+			autoDirty();
+		}
+		return this;
+	}
 
 	/**
 	 * Gets the z render depth for this 3-d item widget
 	 * @return depth
 	 */
-	public int getDepth();
+	public int getDepth() {
+		return depth;
+	}
 
 	@Override
-	public ItemWidget setWidth(int width);
+	public ItemWidget setHeight(int height) {
+		super.setHeight(height);
+		return this;
+	}
 
 	@Override
-	public ItemWidget setHeight(int height);
+	public ItemWidget setWidth(int width) {
+		super.setWidth(width);
+		return this;
+	}
+
+	@Override
+	public WidgetType getType() {
+		return WidgetType.ItemWidget;
+	}
+
+	@Override
+	public ItemWidget copy() {
+		return ((ItemWidget) super.copy()).setTypeId(getTypeId()).setData(getData()).setDepth(getDepth());
+	}
 }
