+++ src/main/java/org/getspout/spoutapi/gui/WidgetType.java
@@ -28,32 +28,32 @@ public class WidgetType {
 	private static HashMap<WidgetType, Integer> lookupClass = new HashMap<WidgetType, Integer>();
 	private static HashMap<Integer, WidgetType> lookupId = new HashMap<Integer, WidgetType>();
 	private static int lastId = 0;
-	public static WidgetType Label = new WidgetType(GenericLabel.class, 0);
+	public static WidgetType Label = new WidgetType(Label.class, 0);
 	public static WidgetType HealthBar = new WidgetType(HealthBar.class, 1);
 	public static WidgetType BubbleBar = new WidgetType(BubbleBar.class, 2);
 	public static WidgetType ChatBar = new WidgetType(ChatBar.class, 3);
 	public static WidgetType ChatTextBox = new WidgetType(ChatTextBox.class, 4);
 	public static WidgetType ArmorBar = new WidgetType(ArmorBar.class, 5);
-	public static WidgetType Texture = new WidgetType(GenericTexture.class, 6);
-	public static WidgetType PopupScreen = new WidgetType(GenericPopup.class, 7);
+	public static WidgetType Texture = new WidgetType(Texture.class, 6);
+	public static WidgetType PopupScreen = new WidgetType(PopupScreen.class, 7);
 	public static WidgetType InGameScreen = new WidgetType(null, 8);
-	public static WidgetType ItemWidget = new WidgetType(GenericItemWidget.class, 9);
-	public static WidgetType Button = new WidgetType(GenericButton.class, 10);
-	public static WidgetType Slider = new WidgetType(GenericSlider.class, 11);
-	public static WidgetType TextField = new WidgetType(GenericTextField.class, 12);
-	public static WidgetType Gradient = new WidgetType(GenericGradient.class, 13);
-	public static WidgetType Container = new WidgetType(GenericContainer.class, 14, true);
-	public static WidgetType EntityWidget = new WidgetType(GenericEntityWidget.class, 15);
-	public static WidgetType OverlayScreen = new WidgetType(GenericOverlayScreen.class, 16);
+	public static WidgetType ItemWidget = new WidgetType(ItemWidget.class, 9);
+	public static WidgetType Button = new WidgetType(Button.class, 10);
+	public static WidgetType Slider = new WidgetType(Slider.class, 11);
+	public static WidgetType TextField = new WidgetType(TextField.class, 12);
+	public static WidgetType Gradient = new WidgetType(Gradient.class, 13);
+	public static WidgetType Container = new WidgetType(Container.class, 14, true);
+	public static WidgetType EntityWidget = new WidgetType(EntityWidget.class, 15);
+	public static WidgetType OverlayScreen = new WidgetType(OverlayScreen.class, 16);
 	public static WidgetType HungerBar = new WidgetType(HungerBar.class, 17);
 	public static WidgetType ExpBar = new WidgetType(ExpBar.class, 18);
-	public static WidgetType CheckBox = new WidgetType(GenericCheckBox.class, 19);
-	public static WidgetType RadioButton = new WidgetType(GenericRadioButton.class, 20);
-	public static WidgetType ListWidget = new WidgetType(GenericListWidget.class, 21);
+	public static WidgetType CheckBox = new WidgetType(CheckBox.class, 19);
+	public static WidgetType RadioButton = new WidgetType(RadioButton.class, 20);
+	public static WidgetType ListWidget = new WidgetType(ListWidget.class, 21);
 	//public static WidgetType DirtBackground = new WidgetType(DirtBackground.class, 22); Only on the client :P
 	//public static WidgetType ScrollArea = new WidgetType(GenericScrollArea.class, 23); Not yet :P
-	public static WidgetType ComboBox = new WidgetType(GenericComboBox.class, 25);
-	public static WidgetType Slot = new WidgetType(GenericSlot.class, 27);
+	public static WidgetType ComboBox = new WidgetType(ComboBox.class, 25);
+	public static WidgetType Slot = new WidgetType(Slot.class, 27);
 	private final int id;
 	private final boolean server;
 	private final Class<? extends Widget> widgetClass;
